회원 기능 합치기 완료
