+++ YAHO/src/main/java/com/njm/yaho/controller/util/EmailAuthController.java
@@ -10,12 +10,17 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.njm.yaho.domain.oracle.user.UserOCDTO;
+import com.njm.yaho.mapper.oracle.user.UserMapperOC;
+import com.njm.yaho.service.user.UserService;
+import com.njm.yaho.service.util.CaptchaService;
 import com.njm.yaho.service.util.EmailAuthService;
+import com.njm.yaho.service.util.MailService;
 
 import jakarta.servlet.http.HttpSession;
 
 @RestController
-@RequestMapping("/auth")
+@RequestMapping("/Auth")
 public class EmailAuthController {
 	@Autowired
     private EmailAuthService emailAuthService;
@@ -33,13 +38,6 @@ public class EmailAuthController {
 	// 인증코드 전송
     @PostMapping("/send")
     public ResponseEntity<String> sendCode(@RequestParam String email) {
-        emailAuthService.sendAuthCode(email);
-        return ResponseEntity.ok("인증 코드가 이메일로 전송되었습니다.");
-    }
-    
-    // 테스트용
-    @GetMapping("/send")
-    public ResponseEntity<String> testSend(@RequestParam String email) {
         emailAuthService.sendAuthCode(email);
         return ResponseEntity.ok("인증 코드가 이메일로 전송되었습니다.");
     }
@@ -70,4 +68,53 @@ public class EmailAuthController {
     	emailAuthService.deleteByEmail(email);  // 또는 서비스 통해 삭제
         return ResponseEntity.ok("인증 정보가 만료되어 삭제되었습니다.");
     }
+    
+    // ====================== 계정 찾기 ========================
+    
+    @Autowired
+	private UserService userService;
+	
+	@Autowired
+	private UserMapperOC userMapperOC;
+	
+	@Autowired
+	private MailService mailService;
+	
+	@Autowired
+	private CaptchaService captchaService;
+    
+    // ID 찾기
+    @GetMapping("/find-id/send")
+	public ResponseEntity<String> sendUserIdToEmail(@RequestParam String email) {
+	    UserOCDTO user = userMapperOC.findByEmail(email);
+	    if (user == null) {
+	        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("존재하지 않는 이메일입니다.");
+	    }
+
+	    // 이메일 전송
+	    mailService.sendUserIdEmail(email, user.getUserId());
+
+	    return ResponseEntity.ok("아이디 전송 완료");
+	}
+
+    // 비밀번호 찾기(초기화)
+	@PostMapping("/find-pw/reset")
+	public ResponseEntity<String> resetPassword(@RequestParam String userId,
+	                                             @RequestParam String email,
+	                                             @RequestParam("captcha") String captchaToken) {
+	    // ✅ 1. reCAPTCHA 검증
+	    if (!captchaService.verifyCaptcha(captchaToken)) {
+	        return ResponseEntity.status(HttpStatus.FORBIDDEN).body("reCAPTCHA 인증 실패");
+	    }
+
+	    try {
+	        boolean result = userService.resetPassword(userId, email);
+	        if (!result) {
+	            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("일치하는 회원 정보가 없습니다.");
+	        }
+	        return ResponseEntity.ok("임시 비밀번호 전송 완료");
+	    } catch (IllegalStateException | IllegalArgumentException e) {
+	        return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body(e.getMessage());
+	    }
+	}
 }
