+++ YAHO/src/main/java/com/njm/yaho/service/user/UserServiceImpl.java
@@ -112,7 +112,7 @@ public class UserServiceImpl implements UserService {
 	        return "/upload/profile/" + newFilename;
 	}
   
-  @Override
+	@Override
 	public boolean resetPassword(String userId, String email) {
 		int todayCount = userMapperOC.countTodayPwReset(userId);
 	    if (todayCount >= 3) {
