+++ YAHO/src/main/resources/templates/User/User_Insert.html
@@ -116,6 +116,7 @@ function send() {
 										<td align=right BORDER="0" HSPACE="0" VSPACE="0"></td>
 									</tr>
 								</table>
+								
 								<table cellpadding=3 cellspacing=1 border=0 width=100%>
 									<tr>
 										<TD BGCOLOR="#EFF4F8">&nbsp;프로필사진<font color=red>&nbsp;*</font></td>
@@ -171,6 +172,7 @@ function send() {
 											size=60 value=""></td>
 									</tr>
 								</table>
+								
 								<table cellpadding=0 cellspacing=0 border=0 width=100%>
 									<tr bgcolor=#7AAAD5>
 										<td valign=bottom><img src="./img/touhou2.jpg" align=left
@@ -186,10 +188,11 @@ function send() {
 								</table>
 							</td>
 						</tr>
-						</td>
-						</tr>
 					</table>
-					</form>
+				</td>
+			</tr>
+		</table>
+	</form>
 
 </body>
 </html>
